Revert "Create a pytest hook"
This reverts commit c7e67859342b2cb79cb22bd878e7cc3ba4920aa3.
